Added QR code to teacher class and event modals.
