--- client/.env.example
@@ -1,11 +1,0 @@
-VITE_FIREBASE_APIKEY=VITE_FIREBASE_APIKEY
-VITE_FIREBASE_AUTHDOMAIN=VITE_FIREBASE_AUTHDOMAIN
-VITE_FIREBASE_PROJECTID=VITE_FIREBASE_PROJECTID
-VITE_FIREBASE_STORAGEBUCKET=VITE_FIREBASE_STORAGEBUCKET
-VITE_FIREBASE_MESSAGINGSENDERID=VITE_FIREBASE_MESSAGINGSENDERID
-VITE_FIREBASE_APPID=VITE_FIREBASE_APPID
-VITE_FIREBASE_MEASUREMENTID=VITE_FIREBASE_MEASUREMENTID
-
-VITE_FRONTEND_HOSTNAME=localhost:3000
-VITE_BACKEND_HOSTNAME=http://localhost:3001
-VITE_ADMIN_EMAIL="something@gmail.com"
