+++ client/src/App.tsx
@@ -25,9 +25,9 @@ import { Login } from "./components/login/Login";
 import { Playground } from "./components/playground/Playground";
 import { Profile } from "./components/profile/Profile";
 import { ProtectedRoute } from "./components/ProtectedRoute";
-import { CheckInHandler } from "./components/qrcode/CheckInHandler";
-import { ClassCheckInHandler } from "./components/qrcode/ClassCheckInHandler";
-import { EventCheckInHandler } from "./components/qrcode/EventCheckInHandler";
+import { CheckInHandler } from "./components/bookings/teacherView/qrcode/CheckInHandler";
+import { ClassCheckInHandler } from "./components/bookings/teacherView/qrcode/ClassCheckInHandler";
+import { EventCheckInHandler } from "./components/bookings/teacherView/qrcode/EventCheckInHandler";
 import { Resources } from "./components/resources/Resources";
 import { Reviews } from "./components/reviews/Reviews";
 import { Settings } from "./components/profile/Settings";
