+++ client/src/components/bookings/teacherView/qrcode/ClassCheckInHandler.jsx
@@ -4,10 +4,10 @@ import { Box, Button, Center, Spinner, Text, VStack } from "@chakra-ui/react";
 
 import { useNavigate, useParams } from "react-router-dom";
 
-import { useAuthContext } from "../../contexts/hooks/useAuthContext";
-import { useBackendContext } from "../../contexts/hooks/useBackendContext";
+import { useAuthContext } from "../../../../contexts/hooks/useAuthContext";
+import { useBackendContext } from "../../../../contexts/hooks/useBackendContext";
 
-export const EventCheckInHandler = () => {
+export const ClassCheckInHandler = () => {
   const { id } = useParams();
   const navigate = useNavigate();
   const { backend } = useBackendContext();
@@ -16,15 +16,22 @@ export const EventCheckInHandler = () => {
   const [loading, setLoading] = useState(true);
   const [error, setError] = useState(null);
   const params = useParams();
+  const hasCheckedIn = useRef(false);
 
   useEffect(() => {
     const handleCheckIn = async () => {
+      // check if user has checked in previously
+      if (hasCheckedIn.currentUser) {
+        return;
+      }
+      hasCheckedIn.currentUser = true;
+
       try {
         if (!currentUser?.uid) {
           const id = params.id;
 
           // removed baseURL, was preventing the redirect to login from happening
-          localStorage.setItem("qrcode_redirect", `/check-in/event/${id}`);
+          localStorage.setItem("qrcode_redirect", `/check-in/class/${id}`);
           // throw new Error("No user ID found");
           navigate("/login");
         }
@@ -34,15 +41,18 @@ export const EventCheckInHandler = () => {
         );
         const studentId = studentResponse.data.id;
 
-        await backend.post("/event-enrollments", {
-          student_id: studentId,
-          event_id: id,
-          // this statement has no effect since attendance defaults to false in backend route (event_enrollments.ts)
-          attendance: true,
-        });
+        // Format current date as YYYY-MM-DD
+        const today = new Date().toISOString().split("T")[0];
 
-        const eventResponse = await backend.get(`/events/${id}`);
-        setTitle(eventResponse.data[0].title);
+        // Class-specific endpoint
+        // await backend.post("/class-enrollments", {
+        //   studentId: studentId,
+        //   classId: id,
+        //   attendance: today,
+        // });
+
+        const classResponse = await backend.get(`/classes/${id}`);
+        setTitle(classResponse.data[0].title);
 
         setLoading(false);
       } catch (err) {
