+++ client/src/components/bookings/teacherView/qrcode/QRCode.jsx
